lxd: Moves swagger response definitions to response package.
